refactor: Use our custom logger for retries
Not knowing a retry happened can make it harder to debug[1][2]. This
adds a log to when retries happen.

Our custom logger was moved to a place that's _in_ the
`knowledge-graph` importable package to facilitate this.

[1]: https://github.com/climatepolicyradar/knowledge-graph/pull/1103
[2]: https://app.prefect.cloud/account/4b1558a0-3c61-4849-8b18-3e97e0516d78/workspace/1753b4f0-6221-4f6a-9233-b146518b4545/runs/flow-run/069ca44f-c25d-7470-8000-e39b20365c7f?preview=true&tab=logs
