+++ flows/utils.py
@@ -2,7 +2,6 @@ import asyncio
 import functools
 import inspect
 import json
-import logging
 import os
 import re
 import textwrap
@@ -55,6 +54,7 @@ from knowledge_graph.cloud import (
     function_to_flow_name,
     generate_deployment_name,
 )
+from knowledge_graph.utils import get_logger
 
 T = TypeVar("T")
 U = TypeVar("U")
@@ -865,30 +865,6 @@ async def gather_and_report(
             )
 
     return results
-
-
-LoggingAdapter = logging.LoggerAdapter[logging.Logger]
-
-
-def get_logger() -> logging.Logger | LoggingAdapter:
-    """
-    Get a logger via Prefect.
-
-    You can overwrite the logging level[2]. If not running in a flow
-    or task run context, a logger that doesn't send to the Prefect API
-    is returned.
-
-    > `get_run_logger()` can only be used in the context of a flow or task.
-    > To use a normal Python logger anywhere with your same configuration, use `get_logger()` from `prefect.logging`.
-    > The logger retrieved with `get_logger()` will not send log records to the Prefect API.
-
-    [1]: https://docs.prefect.io/v3/how-to-guides/workflows/add-logging
-    [2]: https://docs.prefect.io/v3/api-ref/settings-ref#logging-level
-    """
-    try:
-        return prefect.logging.get_run_logger()
-    except prefect.exceptions.MissingContextError:
-        return prefect.logging.get_logger()
 
 
 def get_run_name() -> None | str:
