+++ knowledge_graph/utils.py
@@ -1,10 +1,38 @@
 """Utility functions for the knowledge_graph package."""
 
 import json
+import logging
 from collections.abc import Generator, Sequence
 from typing import TypeVar
 
+import prefect
+import prefect.exceptions
+import prefect.logging
 from pydantic import BaseModel, ValidationError
+
+LoggingAdapter = logging.LoggerAdapter[logging.Logger]
+
+
+def get_logger() -> logging.Logger | LoggingAdapter:
+    """
+    Get a logger via Prefect.
+
+    You can overwrite the logging level[2]. If not running in a flow
+    or task run context, a logger that doesn't send to the Prefect API
+    is returned.
+
+    > `get_run_logger()` can only be used in the context of a flow or task.
+    > To use a normal Python logger anywhere with your same configuration, use `get_logger()` from `prefect.logging`.
+    > The logger retrieved with `get_logger()` will not send log records to the Prefect API.
+
+    [1]: https://docs.prefect.io/v3/how-to-guides/workflows/add-logging
+    [2]: https://docs.prefect.io/v3/api-ref/settings-ref#logging-level
+    """
+    try:
+        return prefect.logging.get_run_logger()
+    except prefect.exceptions.MissingContextError:
+        return prefect.logging.get_logger()
+
 
 T = TypeVar("T")
